release v1.1.0
Updated laravel version from 8.40+ t0 8.75+
Updated npm packages.
Started the process to separate all language string from logic, and move them to a separate language file to help with localization.
Implemented a workaround fix to issues that is occurred during first time installation, this work around is temporary and will be improved in a later update.
A new feature was added to display a full list of licenses and attributes for packages whom were used to develop this web app, and any third party artwork.